fix(formation): fix nested-form SSR bug, relocate component (#1962)
Full-branch-sweep findings:

- project-picker.component.html rendered a <form> nested inside
  propose.component.html's outer <form>. Nested <form> elements are
  invalid HTML — the browser's SSR-parsed markup drops the inner one,
  mismatching the client-side render (where the DOM API tolerates the
  nesting) and forcing Angular to destroy and rebuild that subtree on
  hydration instead of hydrating it. Also meant Enter in the parent-
  project search box could trigger the *outer* form's implicit submit
  before hydration replaced the element. [formGroup] binds to any
  element, not just <form>, so swap to a <div> with no functional loss.
- Move WhatsNextPanelComponent from propose/components/ to
  modules/formations/components/ — it's consumed by both propose and
  propose-confirmation, so it belongs at the module level per
  component-architecture.md, not nested under one of its two callers'
  route folders (git mv, so blame history follows the file).
