+++ apps/lfx-one/src/app/modules/formations/components/whats-next-panel/whats-next-panel.component.ts
